requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
Flask-WTF==1.2.1
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
